+++ pycantonese/tests/test_tipa.py
@@ -1,5 +1,5 @@
 from pycantonese.jyutping.parse_jyutping import ONSETS, TONES
-from pycantonese.jyutping.tipa import ONSETS_TIPA, TONES_TIPA, jyutping2tipa
+from pycantonese.jyutping.tipa import ONSETS_TIPA, TONES_TIPA, jyutping_to_tipa
 
 
 # TODO def test_correct_nucleus_set():
@@ -15,8 +15,8 @@ def test_correct_tone_set():
 
 
 def test_jyutping2tipa():
-    assert jyutping2tipa("m4goi1") == ["\\s{m}21", "kOY55"]
-    assert jyutping2tipa("gwong2dung1waa2") == [
+    assert jyutping_to_tipa("m4goi1") == ["\\s{m}21", "kOY55"]
+    assert jyutping_to_tipa("gwong2dung1waa2") == [
         "k\\super w ON25",
         "tUN55",
         "wa25",